revamp folders structure; fix failing tests
