+++ src/test/java/Config/Data.properties
@@ -3,7 +3,7 @@
 ZIPPO_API_BASEURL=http://zippopotam.us/
 ZIPPO_INDIA_58_ENDPOINT=in/110058
 ZIPPO_INDIA_58_COUNTRY=India
-ZIPPO_INDIA_58_JSON_SCHEMA_PATH=JsonSchemas/Zippo_India_110058_Json_Schema.json
+ZIPPO_INDIA_58_JSON_SCHEMA_PATH=jsonSchemas/Zippo_India_110058_Json_Schema.json
 ZIPPO_INDIA_58_KEYS_COUNT_IN_RESPONSE=4
 ZIPPO_INDIA_58_PLACENAME_COUNT_IN_RESPONSE=5
 
@@ -13,7 +13,7 @@ ZIPPO_INDIA_58_PLACENAME_COUNT_IN_RESPONSE=5
 
 ZIPPO_US_90210_ENDPOINT=us/90210
 ZIPPO_US_90210_COUNTRY=United States
-ZIPPO_US_90210_JSON_SCHEMA_PATH=JsonSchemas/Zippo_US_90210_Json_Schema.json
+ZIPPO_US_90210_JSON_SCHEMA_PATH=jsonSchemas/Zippo_US_90210_Json_Schema.json
 ZIPPO_US_90210_KEYS_COUNT_IN_RESPONSE=4
 ZIPPO_US_90210_PLACENAME_RESPONSE=Beverly Hills
 
@@ -22,7 +22,7 @@ ZIPPO_US_90210_PLACENAME_RESPONSE=Beverly Hills
 
 
 ZIPPO_MOCK_INDIA_110064_COUNTRY=India
-ZIPPO_MOCK_INDIA_110064_JSON_SCHEMA_PATH=JsonSchemas/Zippo_Mocked_India_110064_Json_Schema.json
+ZIPPO_MOCK_INDIA_110064_JSON_SCHEMA_PATH=jsonSchemas/Zippo_Mocked_India_110064_Json_Schema.json
 ZIPPO_MOCK_INDIA_110064_KEYS_COUNT_IN_RESPONSE=4
 ZIPPO_MOCK_INDIA_110064_PLACENAME_RESPONSE=Mayapuri
