--- src/test/java/Utility/Data.properties
@@ -1,40 +1,0 @@
-## Details related to India_110058 test -->>
-
-ZIPPO_API_BASEURL=http://zippopotam.us/
-ZIPPO_INDIA_58_ENDPOINT=in/110058
-ZIPPO_INDIA_58_COUNTRY=India
-ZIPPO_INDIA_58_JSON_SCHEMA_PATH=JsonSchemas/Zippo_India_110058_Json_Schema.json
-ZIPPO_INDIA_58_KEYS_COUNT_IN_RESPONSE=4
-ZIPPO_INDIA_58_PLACENAME_COUNT_IN_RESPONSE=5
-
-
-
-## Details related to US_90210 test -->>
-
-ZIPPO_US_90210_ENDPOINT=us/90210
-ZIPPO_US_90210_COUNTRY=United States
-ZIPPO_US_90210_JSON_SCHEMA_PATH=JsonSchemas/Zippo_US_90210_Json_Schema.json
-ZIPPO_US_90210_KEYS_COUNT_IN_RESPONSE=4
-ZIPPO_US_90210_PLACENAME_RESPONSE=Beverly Hills
-
-
-## Details related to Mocked_India_110064 test -->>
-
-
-ZIPPO_MOCK_INDIA_110064_COUNTRY=India
-ZIPPO_MOCK_INDIA_110064_JSON_SCHEMA_PATH=JsonSchemas/Zippo_Mocked_India_110064_Json_Schema.json
-ZIPPO_MOCK_INDIA_110064_KEYS_COUNT_IN_RESPONSE=4
-ZIPPO_MOCK_INDIA_110064_PLACENAME_RESPONSE=Mayapuri
-
-
-## Mocked api endpoints --
-postEndPoint1 = /postApi/mockedEndpoint
-postEndPoint2 = /postApi/mockedEndpointByNavpreet
-
-
-
-
-
-## Api Mocked Response Body --
-POST_API_RESPONSE_BODY1_ENDPOINT1 = {"post code": "110064", "country": "India", "country abbreviation": "IN", "places": [{"place name": "Mayapuri", "longitude": "77.1726", "state": "New Delhi", "state abbreviation": "DL", "latitude": "28.6488"}]}
-POST_API_RESPONSE_BODY2_ENDPOINT1 = {"postcode": "110064", "country": "India", "cityName": "New Delhi", "places": "Garden"}
