+++ src/test/java/Utility/WireMockSetup.java
@@ -1,5 +1,7 @@
 package Utility;
 
+import Base.RestUtil;
+import PojoClasses.Address;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import io.restassured.response.Response;
 import org.testng.annotations.AfterClass;
