+++ src/test/resources/jsonSchemas/Zippo_India_110058_Json_Schema.json
