+++ src/test/resources/jsonSchemas/Zippo_Mocked_India_110064_Json_Schema.json
