+++ src/test/resources/jsonSchemas/Zippo_US_90210_Json_Schema.json
